refactor: return fs.Type on err and support all Unix
